+++ tests/tests_unit/test_api/test_signatures.py
@@ -3,7 +3,7 @@ import json
 
 import pytest
 
-from cognite.client._api import assets, data_sets, events, files, sequences
+from cognite.client._api import assets, data_sets, events, files, sequences, time_series
 
 
 class TestListAndIterSignatures:
@@ -38,6 +38,11 @@ class TestListAndIterSignatures:
                 ["root_asset_external_ids", "data_set_external_ids", "asset_subtree_external_ids"],
             ),
             (sequences.SequencesAPI, sequences.SequenceFilter, ["asset_subtree_external_ids", "data_set_external_ids"]),
+            (
+                time_series.TimeSeriesAPI,
+                time_series.TimeSeriesFilter,
+                ["asset_subtree_external_ids", "data_set_external_ids", "include_metadata", "partitions"],
+            ),
             (data_sets.DataSetsAPI, data_sets.DataSetFilter, []),
         ],
     )