+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Changes are grouped as follows
 ### Added
 - DataSets API and support for this in assets, events, time series, files and sequences.
 - .asset helper function on time series.
+- asset external id filter on time series.
 
 ## [1.4.13] - 2020-03-03
 ### Added